Update: Move fili fact serializer to client
